requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
blis==0.7.4
boto3==1.18.62
botocore==1.21.62
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
enum-compat==0.0.3
gitdb==4.0.7
GitPython==3.1.24
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-tabnine==1.2.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.1.0/pt_core_news_lg-3.1.0-py3-none-any.whl
pyarrow==5.0.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pywinpty==1.1.4
pyzmq==22.3.0
regex==2021.10.8
requests==2.26.0
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
streamlit==0.69.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.10
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.1
